pyproject.toml
[project]
name = "wb_api_manager"
version = "0.1.0"
description = "api key, cards list"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "logger>=1.4",
    "logging>=0.4.9.6",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.1",
    "sqlalchemy>=2.0.43",
    "typing>=3.10.0.0",
    "utils>=1.0.2",
]
